feat: update pdg commands
