+++ programs/pdg/write.ts
@@ -5,7 +5,6 @@ import { getAccount } from 'providers';
 import {
   getPredepositGuaranteeContract,
   getStakingVaultContract,
-  getBLSHarnessContract,
 } from 'contracts';
 import {
   callWriteMethodWithReceipt,
@@ -17,10 +16,8 @@ import {
   stringToBigIntArray,
   etherToWei,
   callReadMethod,
-  computeDepositDataRoot,
   logResult,
   logInfo,
-  logError,
   confirmOperation,
   confirmMakeProof,
   getCommandsJson,
@@ -107,122 +104,6 @@ pdgWrite
         deposits,
         depositsY,
       ]);
-    },
-  );
-
-pdgWrite
-  .command('verify-predeposit-bls')
-  .aliases(['verify-bls'])
-  .description('Verifies BLS signature of the deposit')
-  .option('-a, --vault <address>', 'vault address')
-  .option('-w, --withdrawalCredentials <hex>', 'withdrawal credentials')
-  .argument('<deposits>', 'deposits', parseDepositArray)
-  .action(
-    async (
-      deposits: Deposit[],
-      options: { vault: Address; withdrawalCredentials: Hex },
-    ) => {
-      const vault = options.vault;
-      let withdrawalCredentials = options.withdrawalCredentials;
-
-      if (!vault && !withdrawalCredentials) {
-        logError('You must provide either vault or withdrawal credentials');
-        return;
-      } else if (vault && withdrawalCredentials) {
-        logError('You can only provide one of vault or withdrawal credentials');
-        return;
-      }
-      const bls = getBLSHarnessContract();
-      const hideMetadataSpinner = showSpinner({
-        type: 'bouncingBar',
-        message: 'Loading metadata...',
-      });
-      const pdg = await getPredepositGuaranteeContract();
-      const PREDEPOSIT_AMOUNT = await callReadMethod(pdg, 'PREDEPOSIT_AMOUNT');
-
-      if (vault) {
-        const vaultContract = getStakingVaultContract(vault);
-        const wc = await callReadMethod(vaultContract, 'withdrawalCredentials');
-        withdrawalCredentials = wc;
-      }
-      hideMetadataSpinner();
-
-      for (const deposit of deposits) {
-        // amount check
-        if (deposit.amount !== PREDEPOSIT_AMOUNT) {
-          logError(
-            `❌ Deposit amount is not equal to PREDEPOSIT_AMOUNT for pubkey ${deposit.pubkey}`,
-          );
-          return;
-        } else {
-          logInfo(`✅ AMOUNT VALID for Pubkey ${deposit.pubkey}`);
-        }
-
-        // depositDataRoot check
-        const depositDataRoot = computeDepositDataRoot(
-          deposit.pubkey,
-          withdrawalCredentials,
-          deposit.signature,
-          deposit.amount,
-        );
-        if (depositDataRoot != deposit.depositDataRoot) {
-          logError(
-            `❌ depositDataRoot does not match ${deposit.pubkey}, actual root: ${depositDataRoot}`,
-          );
-          return;
-        } else {
-          logInfo(`✅ depositDataRoot VALID for Pubkey ${deposit.pubkey}`);
-        }
-
-        // local BLS check
-        const isBLSValid = isValidBLSDeposit(deposit, withdrawalCredentials);
-        if (!isBLSValid) {
-          logError(
-            `❌ Offchain - BLS signature is not valid for Pubkey ${deposit.pubkey}`,
-          );
-          return;
-        } else {
-          logInfo(`✅ SIGNATURE VALID for Pubkey ${deposit.pubkey}`);
-        }
-
-        // onchain BLS check
-        const {
-          pubkeyY_a,
-          pubkeyY_b,
-          sigY_c0_a,
-          sigY_c0_b,
-          sigY_c1_a,
-          sigY_c1_b,
-        } = expandBLSSignature(deposit.signature, deposit.pubkey);
-        const hideSpinner = showSpinner({
-          type: 'bouncingBar',
-          message: 'Checking onchain againts BLSHarness contract',
-        });
-        try {
-          await bls.read.verifyDepositMessage([
-            deposit,
-            {
-              pubkeyY: { a: pubkeyY_a, b: pubkeyY_b },
-              signatureY: {
-                c0_a: sigY_c0_a,
-                c0_b: sigY_c0_b,
-                c1_a: sigY_c1_a,
-                c1_b: sigY_c1_b,
-              },
-            },
-            withdrawalCredentials,
-          ]);
-
-          hideSpinner();
-          logInfo(`✅ ONCHAIN 🔗 SIGNATURE VALID for Pubkey ${deposit.pubkey}`);
-        } catch (err) {
-          hideSpinner();
-          logError(
-            `❌ Onchain - BLS signature is not valid for Pubkey ${deposit.pubkey}`,
-          );
-          logError(err);
-        }
-      }
     },
   );
